Fetch article in a separate function
